Simplify service management
This simplification takes advantage of the fact that ensure => running
is equivalent to ensure => true (and similar for stopped and false) to
reduce the number of variables we have to track to ensure a service is
both enabled and running or disabled and stopped.

This also fixes krb5-kpropd to run on boot without waiting for a puppet
pulse.

Change-Id: I7255cd6bd719946b7525bcbbff983038d81b48a6